Additional Context supplied to model
In this build, the request sent to the claude model includes
other example marked exam submissions, accessed vai the database.

This leads to an increase in cost but it was hoped that additional context could improve the accuracy of results.

This commit exists for posterity to demonstrate the approach taken to generate these results.